build: move main files out of examples
